Created new base classes for search and connector services
