+++ libs/services/designer-client-services/src/lib/base/index.ts
@@ -1,2 +1,10 @@
+// Search
+export { BaseSearchService, getClientBuiltInOperations, getClientBuiltInConnectors } from './search';
+export type { BaseSearchServiceOptions } from './search';
+// Manifests
 export { BaseOperationManifestService, foreachOperationInfo } from './operationmanifest';
 export type { BaseOperationManifestServiceOptions } from './operationmanifest';
+export { frequencyValues } from './manifests/schedule';
+// Connector
+export { BaseConnectorService } from './connector';
+export type { BaseConnectorServiceOptions } from './connector';
