+++ libs/services/designer-client-services/src/lib/base/operationmanifest.ts
@@ -165,6 +165,7 @@ const supportedManifestTypes = [
 export type getAccessTokenType = () => Promise<string>;
 
 export interface BaseOperationManifestServiceOptions {
+  [x: string]: any;
   apiVersion: string;
   baseUrl: string;
   httpClient: IHttpClient;
