+++ libs/services/designer-client-services/src/lib/consumption/connector.ts
@@ -1,0 +1,3 @@
+import { BaseConnectorService } from '../base';
+
+export class ConsumptionConnectorService extends BaseConnectorService {}
