+++ libs/services/designer-client-services/src/lib/standard/index.ts
@@ -4,4 +4,3 @@ export { StandardGatewayService } from './gateway';
 export { StandardOperationManifestService } from './operationmanifest';
 export { StandardOAuthService } from './oAuth';
 export { StandardSearchService } from './search';
-export { frequencyValues } from '../base/manifests/schedule';
